feat: unfiltered-events to read all tty sequences
`espflash(1)`[^1] is using crossterm to implement a `screen(1)`-like
ptty proxy in its "monitor" mode.

`noline`[^2] is a rust crate that implements a `readline(3)`-like line
buffer in a `no_std` environment, like the kinds of chips that
`espflash` targets.

On startup, `noline` does the usual terminal size discovery dance of
moving the cursor "far away" and then querying for the position: the
response to which `crossterm` intercepted and refused to furnish to
`espflash` so it could proxy the response back to the device running
`noline`.

So at this snapshot, `crossterm` has grown a cargo feature that
furnishes the raw bytes of the response as part of its usual input
loop. The alternative would seem to be teaching espflash to pre-parse
the ANSI sequences to "notice" the cursor position request and call the
higher-level crossterm API; adding another ANSI parser/state machine to
the mix adds relatively more steps than "un-parsing" the
`InternalEvent`s, as here.

[^1]: https://github.com/esp-rs/espflash/
[^2]: https://github.com/rustne-kretser/noline